requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1  # 缓存支持

# Data Processing
pandas==2.1.3
numpy==1.26.2
yfinance==0.2.33
scipy==1.11.4  # 高级数学计算
ta==0.11.0  # 技术指标库

# Financial Analysis
numpy-financial==1.0.0  # IRR计算
pyfolio==0.9.2  # 投资组合分析
quantstats==0.0.62  # 量化统计

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Utilities
aiofiles==23.2.1
httpx==0.25.2
python-dateutil==2.8.2
apscheduler==3.10.4  # 定时任务

# Visualization and Export
matplotlib==3.8.2  # 图表生成
plotly==5.18.0  # 交互式图表
openpyxl==3.1.2  # Excel导出
reportlab==4.0.7  # PDF生成
Pillow==10.1.0  # 图像处理

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# API Integration
ccxt==4.1.22  # 加密货币交易所API（可选）
alpha-vantage==2.3.1  # Alpha Vantage API（备选数据源）

# Machine Learning (可选，用于未来功能)
# scikit-learn==1.3.2
# xgboost==2.0.2
# lightgbm==4.1.0